Codestyle refactoring
