+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/presentation/settings/SettingsPresenter.java
@@ -8,7 +8,7 @@ import com.arellomobile.mvp.MvpPresenter;
 import sasd97.java_blog.xyz.yandexweather.R;
 import sasd97.java_blog.xyz.yandexweather.domain.converters.ConvertersConfig;
 import sasd97.java_blog.xyz.yandexweather.domain.settings.SettingsInteractor;
-import sasd97.java_blog.xyz.yandexweather.utils.RxSchedulersAbs;
+import sasd97.java_blog.xyz.yandexweather.utils.RxSchedulers;
 
 /**
  * Created by alexander on 15/07/2017.
@@ -17,10 +17,10 @@ import sasd97.java_blog.xyz.yandexweather.utils.RxSchedulersAbs;
 @InjectViewState
 public class SettingsPresenter extends MvpPresenter<SettingsView> {
 
-    private RxSchedulersAbs schedulers;
+    private RxSchedulers schedulers;
     private SettingsInteractor interactor;
 
-    public SettingsPresenter(@NonNull RxSchedulersAbs schedulers,
+    public SettingsPresenter(@NonNull RxSchedulers schedulers,
                              @NonNull SettingsInteractor interactor) {
         this.schedulers = schedulers;
         this.interactor = interactor;
