+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/utils/RxSchedulersImpl.java
@@ -1,0 +1,28 @@
+package sasd97.java_blog.xyz.yandexweather.utils;
+
+import io.reactivex.Scheduler;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.schedulers.Schedulers;
+
+/**
+ * Created by alexander on 07/07/2017.
+ */
+
+public class RxSchedulersImpl extends RxSchedulers {
+
+    @Override
+    public Scheduler getMainThreadScheduler() {
+        return AndroidSchedulers.mainThread();
+    }
+
+    @Override
+    public Scheduler getIoScheduler() {
+        return Schedulers.io();
+    }
+
+    @Override
+    public Scheduler getComputationScheduler() {
+        return Schedulers.computation();
+    }
+
+}
